fix(lint): split routes/http.rs family under the 600-line gate
http.rs (637), http_tests.rs (613), and http_listener_tests.rs (658) had
crept past the repo's 600-line pre-push gate. Splits machine-identity and
user-prompt settings routes out of http.rs into http_settings_routes.rs,
their tests out of http_tests.rs into http_settings_routes_tests.rs, and
the global-lock endpoint tests out of http_listener_tests.rs into
http_listener_lock_tests.rs. No behavior or test-body changes.
